Move methods into BitOps.other
